refactor: move nodeReadableToWebStream helper to utils
Keep src/index.js focused on the wrapper logic and put the
node-readable -> web-stream adapter alongside the other stream/response
helpers in src/utils.js.
